angolan real estate committee

+++ src/BSetter.sol
@@ -36,20 +36,21 @@ contract BSetter is PoolToken, BStorage {
 
     /// @notice Emitted when a new adjustment speed is set.
     /// @param newAdjustSpeed The new adjustment speed.
-    event NewAdjustSpeed(uint newAdjustSpeed);
+    event NewAdjustSpeed(uint256 newAdjustSpeed);
 
     /// @notice Emitted when a new borrow tracker contract is set.
     /// @param newBorrowTracker New borrow tracker contract.
     event NewBorrowTracker(address newBorrowTracker);
 
-    // called once by the factory at time of deployment
+    /// @notice called once by the factory at time of deployment
     function _initialize(
         string calldata _name,
         string calldata _symbol,
         address _underlying,
         address _collateral
     ) external {
-        _require(msg.sender == factory, Errors.UNAUTHORIZED_CALL); // sufficient check
+        /// @dev sufficient check
+        _require(msg.sender == factory, Errors.UNAUTHORIZED_CALL);
         _setName(_name, _symbol);
         underlying = _underlying;
         collateral = _collateral;
@@ -99,4 +100,3 @@ contract BSetter is PoolToken, BStorage {
         );
     }
 }
-
